added explicit cast from double to integer for retun value of ceil function.

+++ src/search/landmarks/landmark_count_heuristic.cc
@@ -83,7 +83,7 @@ int LandmarkCountHeuristic::get_heuristic_value(const State &state) {
 
     if (use_cost_sharing) {
         double h_val = lm_cost_assignment->cost_sharing_h_value();
-        h = ceil(h_val - epsilon);
+        h = (int)ceil(h_val - epsilon);
     } else {
         lgraph.count_costs();
